+++ src/extension.ts
@@ -1,15 +1,13 @@
 import { ChildProcess, fork } from "child_process";
-import * as path from "path";
-import * as vscode from "vscode";
+import { join } from "path";
+import { commands, ExtensionContext, ProgressLocation, TextDocumentContentProvider, Uri, ViewColumn, window, workspace, WorkspaceFolder } from "vscode";
 
-const previewUri = vscode.Uri.parse("ungit://view");
-const modulePath = path.join(__dirname, "..", "..", "node_modules", "ungit", "bin", "ungit");
+const modulePath = join(__dirname, "..", "..", "node_modules", "ungit", "bin", "ungit");
 let child: ChildProcess;
 
-export class TextDocumentContentProvider implements vscode.TextDocumentContentProvider {
-    public provideTextDocumentContent(uri: vscode.Uri): string {
-        const location = vscode.workspace.rootPath || "";
-        const url = `http://localhost:8448/#/repository?path=${location}`;
+export class UngitTextDocumentContentProvider implements TextDocumentContentProvider {
+    public provideTextDocumentContent(uri: Uri): string {
+        const url = `http://localhost:8448/#/repository?path=${uri.fsPath}`;
         return `
         <div style="position: fixed; height: 100%; width: 100%; margin-left: -20px;">
             <iframe src="${url}" style="border: none;" height="100%" width="100%"></iframe>
@@ -18,9 +16,38 @@ export class TextDocumentContentProvider implements vscode.TextDocumentContentPr
     }
 }
 
+/**
+ * Determine the Ungit root from the current active file.
+ * Let the user pick a workspace if there is no open file.
+ */
 function executeCommand(): void {
-    vscode.window.withProgress({
-        location: vscode.ProgressLocation.Notification,
+    const activeTextEditor = window.activeTextEditor;
+    let workspaceFolderPromise: Thenable<WorkspaceFolder | undefined>;
+    if (activeTextEditor && activeTextEditor.document.uri.scheme === "file") {
+        workspaceFolderPromise = Promise.resolve(workspace.getWorkspaceFolder(activeTextEditor.document.uri));
+    } else if (workspace.workspaceFolders && workspace.workspaceFolders.length === 1) {
+        workspaceFolderPromise = Promise.resolve(workspace.workspaceFolders[0]);
+    } else {
+        workspaceFolderPromise = window.showWorkspaceFolderPick();
+    }
+    workspaceFolderPromise.then((workspaceFolder) => {
+        if (typeof workspaceFolder === "undefined") {
+            window.showErrorMessage("Can't open Ungit outside a workspace.");
+        } else if (workspaceFolder.uri.scheme !== "file") {
+            window.showErrorMessage("Can't open Ungit on remote workpaces.");
+        } else {
+            openInWorkspace(workspaceFolder);
+        }
+    }, () => {
+        window.showErrorMessage("Can't open Ungit: Unable to determine workspace.");
+    });
+}
+
+function openInWorkspace(workspaceFolder: WorkspaceFolder): void {
+    const ungitUri = workspaceFolder.uri.with({scheme: "ungit"});
+    const ungitTabTitle = `Ungit - ${workspaceFolder.name}`;
+    window.withProgress({
+        location: ProgressLocation.Notification,
         title: "Starting Ungit",
         cancellable: true,
     }, (progress, token) => {
@@ -34,15 +61,16 @@ function executeCommand(): void {
             child.stdout.on("data", (message: Buffer) => {
                 const started =
                     (message.toString().includes("## Ungit started ##")) ||
-                    (message.toString().includes("Ungit server already running"));
+                    (message.toString().includes("Ungit server already running")) ||
+                    (message.toString().includes("Error: listen EADDRINUSE 127.0.0.1:8448"));
                 if (started) {
                     progress.report({
                         increment: 100,
                     });
-                    vscode.commands.executeCommand("vscode.previewHtml", previewUri, vscode.ViewColumn.Two, "Ungit").then(() => {
+                    commands.executeCommand("vscode.previewHtml", ungitUri, ViewColumn.Two, ungitTabTitle).then(() => {
                         resolve();
                     }, (reason: string) => {
-                        vscode.window.showErrorMessage(reason);
+                        window.showErrorMessage(reason);
                         reject();
                     });
                 }
@@ -51,10 +79,10 @@ function executeCommand(): void {
     });
 }
 
-export function activate(context: vscode.ExtensionContext): void {
-    const provider = new TextDocumentContentProvider();
-    const registration = vscode.workspace.registerTextDocumentContentProvider("ungit", provider);
-    const disposable = vscode.commands.registerCommand("extension.ungit", executeCommand);
+export function activate(context: ExtensionContext): void {
+    const provider = new UngitTextDocumentContentProvider();
+    const registration = workspace.registerTextDocumentContentProvider("ungit", provider);
+    const disposable = commands.registerCommand("extension.ungit", executeCommand);
     context.subscriptions.push(disposable, registration);
 }