+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Support for Multi Root Workspace
+
 ### Changed
 - Updated Ungit to 1.4.31